EBBR: Use emission animation at night.
Some tweaks have been applied to some buildings as this was needed to
accomplish the emission animation.

On the way of this, found out that the stg file was pointing to some
unavailable xml file. Fixed this and i discover a faster scenery load
time.

This fixes a part of issue: https://github.com/dvanmosselbeen/flightgear-belgian-custom-scenery/issues/68