EBFN: Major update /changes to this mil airfield. Part 1
The old generic EBFN buildings have been replaced by objects of EBBL.
Making thus EBFN with an imaginary setup and it's absolutely not based
on real life fact as like EBBL is.

The airport is still not finished. I still want to place some more
buildings of EBBL on EBFN and finally add the military shared models. On
the end, a huge amount of trees will be set. Somethings with a low
respectable amount of trees like done at EBBL.

The goal was to make a bloateware imaginary military airfield bassed on
objects of EBBL and generic shared models. Like the state now, it was
meant to push the EBBL objects to the shared models instead of
duplicating these objects. But my fresh new Nvidia GeForce RTX 2080 of
my not even 2 weeks old computer is dying and got no choice to avoid
data lose. But in the long run, these objects should moved to the shared
modes as i'm planning to do the same thing with for example EBFS
(Florennes).